+++ NEWS.md
@@ -1,3 +1,11 @@
+# waiter 0.2.2
+
+- Remove `console.log`
+- Add `image` argument to the waiter, allows using a background
+image on the waiter.
+- Fix `hide_on_render` [#94](https://github.com/JohnCoene/waiter/issues/94)
+- Fix full screen not covering entirety of body [#95](https://github.com/JohnCoene/waiter/issues/95)
+
 # waiter 0.2.1
 
 - Fixed `waiter_hide_on_render` see [#79](https://github.com/JohnCoene/waiter/issues/79).